evmap

Function with_meta

pub fn with_meta<K, V, M>(
    meta: M,
) -> (ReadHandle<K, V, M, RandomState>, WriteHandle<K, V, M, RandomState>)
where
    K: Eq + Hash + Clone,
    V: Eq + Hash + ShallowCopy,
    M: 'static + Clone,

Create an empty eventually consistent map with meta information.

Use the Options builder for more control over initialization.