Module evmap::shallow_copy

Types that can be cheaply aliased.

Structs

CopyValue
If you are willing to have your values be copied between the two views of the evmap, wrap them in this type.

Traits

ShallowCopy
Types that implement this trait can be cheaply copied by (potentially) aliasing the data they contain. Only the last shallow copy will be dropped -- all others will be silently leaked (with mem::forget).