Trait evmap::shallow_copy::ShallowCopy

pub trait ShallowCopy {
    unsafe fn shallow_copy(&mut self) -> Self;
}

Types that implement this trait can be cheaply copied by (potentially) aliasing the data they contain. Only the last shallow copy will be dropped -- all others will be silently leaked (with mem::forget).

Required methods

unsafe fn shallow_copy(&mut self) -> Self

Perform an aliasing copy of this value.

The use of this method is only safe if the values involved are never mutated, and only one of the copies is dropped; the remaining copies must be forgotten with mem::forget.

Loading content...