Trait evmap::shallow_copy::ShallowCopy

pub trait ShallowCopy {
    unsafe fn shallow_copy(&self) -> ManuallyDrop<Self>;
}

Types that implement this trait can be cheaply copied by (potentially) aliasing the data they contain. Only the last shallow copy will be dropped – all others will be silently leaked (with mem::forget).

To implement this trait for your own Copy type, write:

use std::mem::ManuallyDrop;

#[derive(Copy, Clone)]
struct T;

impl ShallowCopy for T {
    unsafe fn shallow_copy(&self) -> ManuallyDrop<Self> {
        ManuallyDrop::new(*self)
    }
}

If you have a non-Copy type, the value returned by shallow_copy should point to the same data as the &mut self, and it should be safe to mem::forget either of the copies as long as the other is dropped normally afterwards.

For complex, non-Copy types, you can place the type behind a wrapper that implements ShallowCopy such as Box or Arc. Alternatively, if your type is made up of types that all implement ShallowCopy, consider using the evmap-derive crate, which contains a derive macro for ShallowCopy. See that crate’s documentation for details.

Required Methods

unsafe fn shallow_copy(&self) -> ManuallyDrop<Self>

Perform an aliasing copy of this value.

Safety

The use of this method is only safe if the values involved are never mutated, and only one of the copies is dropped; the remaining copies must be forgotten with mem::forget.