Trait evmap::shallow_copy::ShallowCopy

pub trait ShallowCopy {
    unsafe fn shallow_copy(&mut self) -> Self;
}

Types that implement this trait can be cheaply copied by (potentially) aliasing the data they contain. Only the last shallow copy will be dropped -- all others will be silently leaked (with mem::forget).

To implement this trait for your own Copy type, write:

#[derive(Copy, Clone)]
struct T;

impl ShallowCopy for T {
    unsafe fn shallow_copy(&mut self) -> Self {
        *self
    }
}

If you have a non-Copy type, the value returned by shallow_copy should point to the same data as the &mut self, and it should be safe to mem::forget either of the copies as long as the other is dropped normally afterwards.

Required methods

unsafe fn shallow_copy(&mut self) -> Self

Perform an aliasing copy of this value.

The use of this method is only safe if the values involved are never mutated, and only one of the copies is dropped; the remaining copies must be forgotten with mem::forget.

Loading content...