Struct evmap::WriteHandle

pub struct WriteHandle<K, V, M = (), S = RandomState> where
    K: Eq + Hash + Clone,
    S: BuildHasher + Clone,
    V: Eq + ShallowCopy,
    M: 'static + Clone,  { /* fields omitted */ }

A handle that may be used to modify the eventually consistent map.

Note that any changes made to the map will not be made visible to readers until refresh() is called.

When the WriteHandle is dropped, the map is immediately (but safely) taken away from all readers, causing all future lookups to return None.

Examples

let x = ('x', 42);

let (r, mut w) = evmap::new();

// the map is uninitialized, so all lookups should return None
assert_eq!(r.get_and(&x.0, |rs| rs.len()), None);

w.refresh();

// after the first refresh, it is empty, but ready
assert_eq!(r.get_and(&x.0, |rs| rs.len()), None);

w.insert(x.0, x);

// it is empty even after an add (we haven't refresh yet)
assert_eq!(r.get_and(&x.0, |rs| rs.len()), None);

w.refresh();

// but after the swap, the record is there!
assert_eq!(r.get_and(&x.0, |rs| rs.len()), Some(1));
assert_eq!(r.get_and(&x.0, |rs| rs.iter().any(|v| v.0 == x.0 && v.1 == x.1)), Some(true));

Methods

impl<K, V, M, S> WriteHandle<K, V, M, S> where
K: Eq + Hash + Clone,
S: BuildHasher + Clone,
V: Eq + ShallowCopy,
M: 'static + Clone,

pub fn refresh(&mut self) -> &mut Self

Refresh the handle used by readers so that pending writes are made visible.

This method needs to wait for all readers to move to the new handle so that it can replay the operational log onto the stale map copy the readers used to use. This can take some time, especially if readers are executing slow operations, or if there are many of them.

pub fn pending(&self) -> &[Operation<K, V>]

Gives the sequence of operations that have not yet been applied.

Note that until the first call to refresh, the sequence of operations is always empty.

let x = ('x', 42);

let (r, mut w) = evmap::new();

// before the first refresh, no oplog is kept
w.refresh();

assert_eq!(w.pending(), &[]);
w.insert(x.0, x);
assert_eq!(w.pending(), &[Operation::Add(x.0, x)]);
w.refresh();
w.remove(x.0, x);
assert_eq!(w.pending(), &[Operation::Remove(x.0, x)]);
w.refresh();
assert_eq!(w.pending(), &[]);

pub fn flush(&mut self) -> &mut Self

Refresh as necessary to ensure that all operations are visible to readers.

WriteHandle::refresh will always wait for old readers to depart and swap the maps. This method will only do so if there are pending operations.

pub fn set_meta(&mut self, meta: M) -> M

Set the metadata.

Will only be visible to readers after the next call to refresh().

pub fn insert(&mut self, k: K, v: V) -> &mut Self

Add the given value to the value-set of the given key.

The updated value-set will only be visible to readers after the next call to refresh().

pub fn update(&mut self, k: K, v: V) -> &mut Self

Replace the value-set of the given key with the given value.

With the smallvec feature enabled, replacing the value will automatically deallocate any heap storage and place the new value back into the SmallVec inline storage. This can improve cache locality for common cases where the value-set is only ever a single element.

See the doc section on this for more information.

The new value will only be visible to readers after the next call to refresh().

pub fn clear(&mut self, k: K) -> &mut Self

Clear the value-set of the given key, without removing it.

If a value-set already exists, this will clear it but leave the allocated memory intact for reuse, or if no associated value-set exists an empty value-set will be created for the given key.

The new value will only be visible to readers after the next call to refresh().

pub fn remove(&mut self, k: K, v: V) -> &mut Self

Remove the given value from the value-set of the given key.

The updated value-set will only be visible to readers after the next call to refresh().

pub fn empty(&mut self, k: K) -> &mut Self

Remove the value-set for the given key.

The value-set will only disappear from readers after the next call to refresh().

pub fn purge(&mut self) -> &mut Self

Purge all value-sets from the map.

The map will only appear empty to readers after the next call to refresh().

Note that this will iterate once over all the keys internally.

pub unsafe fn retain<F>(&mut self, k: K, f: F) -> &mut Self where
F: FnMut(&V, bool) -> bool + 'static + Send + Sync,

Retain elements for the given key using the provided predicate function.

The remaining value-set will only be visible to readers after the next call to refresh()

Note that the given closure is called twice for each element, once when called, and once on swap. It must retain the same elements each time, otherwise a value may exist in one map, but not the other, leaving the two maps permanently out-of-sync. This is really bad, as values are aliased between the maps, and are assumed safe to free when they leave the map during a refresh. Returning true when retain is first called for a value, and false the second time would free the value, but leave an aliased pointer to it in the other side of the map.

The arguments to the predicate function are the current value in the value-set, and true if this is the first value in the value-set on the second map, or false otherwise. Use the second argument to know when to reset any closure-local state to ensure deterministic operation.

So, stated plainly, the given closure must return the same order of true/false for each of the two iterations over the value-set. That is, the sequence of returned booleans before the second argument is true must be exactly equal to the sequence of returned booleans at and beyond when the second argument is true.

pub fn fit(&mut self, k: K) -> &mut Self

Shrinks a value-set to it's minimum necessary size, freeing memory and potentially improving cache locality by switching to inline storage if the smallvec feature is used.

The optimized value-set will only be visible to readers after the next call to refresh()

pub fn fit_all(&mut self) -> &mut Self

Like WriteHandle::fit, but shrinks all value-sets in the map.

The optimized value-sets will only be visible to readers after the next call to refresh()

pub fn reserve(&mut self, k: K, additional: usize) -> &mut Self

Reserves capacity for some number of additional elements in a value-set, or creates an empty value-set for this key with the given capacity if it doesn't already exist.

Readers are unaffected by this operation, but it can improve performance by pre-allocating space for large value-sets.